Add files via upload
The "SCAMP To Scale" and the "Electronics Box v1" are the CAD files of what we believe we need for SCAMP (not including the soft and hard stops - to be determined later, and camera(s)). The other files are referenced within the assemblies.  These are all what we believe to be accurate first draft of SCAMP.